requirements.txt
python-telegram-bot==20.7
firebase-admin
Flask
gunicorn
python-dotenv
